fix(base): move class FileLoader from Entity directory to Event
Entity directory is reserved only for describing data from DB

Resovle BAM-517
