+++ src/Centreon/ServiceProvider.php
@@ -36,14 +36,13 @@
 
 namespace Centreon;
 
-use Centreon\Infrastructure\Event\EventDispatcher;
-use Centreon\Domain\Entity\FileLoader;
 use Pimple\Container;
 use Pimple\Psr11\ServiceLocator;
 use CentreonLegacy\ServiceProvider as LegacyServiceProvider;
 use Centreon\Application\Webservice;
 use Centreon\Infrastructure\Provider\AutoloadServiceProviderInterface;
 use Centreon\Infrastructure\Service;
+use Centreon\Infrastructure\Event;
 use Centreon\Infrastructure\Service\CentreonWebserviceService;
 use Centreon\Infrastructure\Service\CentreonClapiService;
 use Centreon\Infrastructure\Service\CentcoreConfigService;
@@ -75,6 +74,7 @@ class ServiceProvider implements AutoloadServiceProviderInterface
     const CENTREON_DB_MANAGER = 'centreon.db-manager';
     const CENTREON_CLAPI = 'centreon.clapi';
     const UPLOAD_MANGER = 'upload.manager';
+    const CENTREON_EVENT_DISPATCHER = 'centreon.event_dispatcher';
     const CENTREON_USER = 'centreon.user';
     const YML_CONFIG = 'yml.config';
 
@@ -208,14 +208,15 @@ class ServiceProvider implements AutoloadServiceProviderInterface
             return $service;
         };
 
-        $pimple['centreon.event_dispatcher'] = function (Container $container) {
-            $eventDispatcher = new EventDispatcher();
+        $pimple[static::CENTREON_EVENT_DISPATCHER] = function (Container $container) : Event\EventDispatcher {
+            $eventDispatcher = new Event\EventDispatcher();
             $eventDispatcher->setDispatcherLoader(
-                new FileLoader(
+                new Event\FileLoader(
                     _CENTREON_PATH_ . '/www/modules/',
                     'custom-module-form.php'
                 )
             );
+
             return $eventDispatcher;
         };
     }
